[REFACTORY] split UI helper exports from component modules
