+++ ui/src/components/status-breakdown.tsx
@@ -14,7 +14,7 @@ import { QueryStatusDistributionResult } from "@/lib/types";
 import { formatTimestampByGranularity } from "@/lib/utils/date-formatting";
 
 import { useQuery } from "@tanstack/react-query";
-import { useDateRange } from "@/contexts/date-range-context";
+import { useDateRange } from "@/contexts/date-range";
 import { getQueryStatusDistribution } from "@/api/queries";
 import { Skeleton } from "@/components/ui/skeleton";
